monitor/client/requirements.txt
grpcio==1.38
grpcio-tools
picamera